Change variable to size_t type.

+++ ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: rde-oliv <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/05 14:07:21 by rde-oliv          #+#    #+#             */
-/*   Updated: 2021/01/05 16:40:02 by rde-oliv         ###   ########.fr       */
+/*   Updated: 2021/01/06 09:26:22 by rde-oliv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int	main(int argc, char **argv)
 		for (int i = 1; i < argc; i++)
 		{
 			str = argv[i];
-			for (int j = 0; j < str.size(); j++)
+			for (size_t j = 0; j < str.size(); j++)
 				std::cout << (char)toupper(str.at(j));
 		}
 	}
